+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.4.0-alpha.2](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.4.0-alpha.1...v1.4.0-alpha.2) (2023-09-11)
+
+
+### Bug Fixes
+
+* cls 指标应该在报告数据出来时就统一单位，而不是等到用到数据时才开始转换，处理时机不对 ([07fc2bb](https://github.com/zhuwei-ones/performance-auto-test/commit/07fc2bb5a676cfe3b5e7eebab45fef5eea7fc5a9))
+
+## [1.4.0-alpha.1](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.4.0-alpha.0...v1.4.0-alpha.1) (2023-09-07)
+
+
+### Features
+
+*  format  hooks performance result ([3c5c0f3](https://github.com/zhuwei-ones/performance-auto-test/commit/3c5c0f37491a986ffffb395e8afe73c35aa7485c))
+
+## [1.4.0-alpha.0](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.3.0...v1.4.0-alpha.0) (2023-09-07)
+
+
+### Features
+
+* support hooks to control if Interrupt the performance test ([a97168c](https://github.com/zhuwei-ones/performance-auto-test/commit/a97168c461db3c7e4dbf5cf8ff9fbf011d45ffe3))
+
+## [1.2.0](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.1.0...v1.2.0) (2023-07-21)
+
+## [1.1.0](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.1.0-alpha.24...v1.1.0) (2023-07-21)
+
 ## [1.3.0](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.1.0-alpha.24...v1.3.0) (2023-07-21)
 
 ## [1.1.0-alpha.24](https://github.com/zhuwei-ones/performance-auto-test/compare/v1.1.0-alpha.23...v1.1.0-alpha.24) (2023-07-18)